requirements.txt
certifi==2017.11.5
chardet==3.0.4
FAdo3==1.0
future==0.16.0
idna==2.6
requests==2.18.4
urllib3==1.22
